modify in delete

+++ src/services/user/deleteUser.service.ts
@@ -4,22 +4,10 @@ import { AppError } from "../../errors/AppError";
 
 
 const deleteUserService = async (
-  id: string,
   userId: string
 ): Promise<void> => {
   const userRepository = AppDataSource.getRepository(User);
-
-  const findUser = await userRepository.findOneBy({id});
-
-  if(!findUser) {
-    throw new AppError(404, "User Not Found");
-  }
-
-  if(userId != findUser.id) {
-    throw new AppError(403, "You cannot delete other users")
-  }
-
-  await userRepository.delete({id})
+  await userRepository.delete({id: userId})
 }
 
 export default deleteUserService;
